Change the interface of DrawLine.
Use Coord as the parameter of constructor.

+++ include/Triangle/Coord.h
@@ -12,13 +12,13 @@ namespace luba {
 class Coord
 {
 public:
-  double x;
-  double y;
-  double z;
+  float x;
+  float y;
+  float z;
 
 public:
   Coord() : x(0.0), y(0.0), z(0.0) {}
-  Coord(double pX, double pY, double pZ) : x(pX), y(pY), z(pZ) {}
+  Coord(float pX, float pY, float pZ) : x(pX), y(pY), z(pZ) {}
 
 };
